requirements.txt
asn1crypto==1.5.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
fsspec==2024.9.0
gunicorn==23.0.0
huggingface-hub==0.25.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.1
packaging==24.1
pandas==2.2.3
pg8000==1.31.2
pillow==10.4.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
scramp==1.4.5
sentence-transformers==3.1.1
setuptools==75.1.0
six==1.16.0
SQLAlchemy==2.0.35
sympy==1.13.3
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.1
tqdm==4.66.5
transformers==4.44.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.4
